Grease Pencil: Improve draw performance at beginning and end of stroke
This avoids re-triangulation of the entire drawing at the beginning and
at the end of drawing a new stroke.

Currently, this only works in the (most common) case when drawing
on top instead of to the back of the drawing.

Instead of tagging the runtime caches for the triangulation
and the plane normals for re-computation, copy the caches
of all the strokes that didn't change and update the cache of
the new stroke immediatly.