Remove properties from keymap if they are already default
This is necessary to allow the operator to get overridden by tool
settings. Otherwise it has no effect.

Note that the split operator does not respond to tool settings in main
currently. This will be fixed in a followup fix PR.

Also note that the toolsettings are mixing props for the two operators,
which should be standardized somehow in said fix.