Fix #151274: Lattice artifacts whith the "Outside" option enabled
Caused by 3e8250e60c .

Not all edges are drawn when the `Outside` (LT_OUTSIDE) option is used.
Still, these are included in `lattice_render_edges_len_get`, so this is
what the Indexbuffer is created with.

Note there is already a comment: "/* TODO: remove internal coords. */"
to improve upon this, but until this is tackled, just fill those gaps
with zeros to hide redundant edges, just as in e.g. 0c41f53cf738 .