Fix #151159: Update issue after switching to scene for the first time
This issue was caused by the fact that `ED_scene_change_update` created
a new depsgraph but didn't activate it.
When execution gets to `pose_channel_flush_to_orig_if_needed`, it
silently fails, because the depsgraph isn't active.
This then causes the bones `pose_mat` to not be copied to the original
(so it remains the outdated one). When selecting the bone
this triggers a `deg_create_eval_copy` which copies the outdated
`pose_mat` to the evaluated state.
This is then fixed by any recomputation of the `pose_mat` (e.g. bone
transformation).

The fix ensures that `ED_scene_change_update` marks the depsgraph active.
This then allows `pose_channel_flush_to_orig_if_needed` to succeed.