Fix #151267: Crash joining specific objects
Exposed by 340f9d7ff3

These "specific" objects had no vertex groups in
`mesh.vertex_group_names` yet did have a `CD_MDEFORMVERT` custom data
layer (with `MDeformVert` that actually have a non-zero `MDeformWeight`
array).
This shouldnt be a valid mesh really, couldnt spot an operation that
removes `mesh.vertex_group_names` but not the `CD_MDEFORMVERT` though
(might be "old" meshes from some time in the past).

If we run into such scenario, a proper `index_map` cannot be built and
an illegal access to it happens.

Note 4.5 was crashing as well, but the order of selection did matter
there in the file from the report (prior to 340f9d7ff3, we would join
into the active mesh, after 340f9d7ff3 we are joining int a new mesh, so
all selected meshes are now iterated over...)

So to resolve, we now consider mesh.vertex_group_names the actual source
of truth (not the existence of the `CD_MDEFORMVERT` layers), if that
gathered list from all participating objects is empty, return early.