Compositor: Add new path options to the File Output node
This patch adds two new path options to the File Output node:

- Separate Directories: Saves each image in its own separate directory
  with the same name as the file, but without the frame numbering. This
  is disabled by default. Example before and after:
  - /tmp/Image001.png
  - /tmp/Image002.png
  - /tmp/Alpha001.png
  - /tmp/Alpha002.png

  - /tmp/Image/Image001.png
  - /tmp/Image/Image002.png
  - /tmp/Alpha/Alpha001.png
  - /tmp/Alpha/Alpha002.png
- File Extension: Adds the image extension to the file name. This is
  enabled by default.

The File Extension is not 100% backward compatible, because the
extension option was previously taken from the scene and attempting to
use that would not be reliable because node groups could be used in many
scenes. In most cases, this option was always enabled, so this should
not have much of an impact in practice.