Fix #151284: Adjust Mouse Cursor as Brush Size Decreases
For painting operations we show both a mouse cursor and a circle that
represents the brush size or influence. As the brush size decreases
these two things can overlap and obscure each other. This PR makes the
mouse cursor change to just a dot when small and to nothing when tiny.