UI: Do Not Join Aligned Area Edges
As we move screen area edges, if any line up exactly then we join them
together so that they act as one. The become a single edge permanently
without a way to break them. This PR removes that functionality. Under
the hood what currently happens is that area corner vertexes are merged
if duplicated. As are area edges. This PR just turns off this
deduplication.