Debug Steps behavior for recording internal solver steps
This behavior causes the XPBD solver node to record data about individual
substeps and constraint iterations during the update process. It is useful for
investigating problems that occur on small scales and are easily hidden by the
cumulative result of the overall solver time step.

The code path for debugging is only used when the Debug Steps bundle exists, so
any overhead during normal use is avoided.

The result of the debug recording is stored in a dedicated bundle. The data
representing incremental steps is stored as a list of geometries, which is not
directly accessible with regular list nodes at this point. For this reason a
dedicated Debug Extract node is added which allows selecting geometry from any
of the substeps and/or constraint iterations.