Fix #151083: Separate by Material breaks custom normals part II
43d1685bc318 already fixed this for editmode, but calling the operator
in objectmode was still wrong.

When in objectmode, ensure we have updated normals for the bmesh, so use
`BMeshFromMeshParams` for that.

NOTE: `BMeshFromMeshParams` `calc_face_normal` seems to enough to fix
this, I havent spent the time to track down in detail where this is
needed in the rabbit hole that normals calculation code is, other code
that relies in updated bmesh normals seems to do both `calc_face_normal`
**and** `calc_vert_normal` though...