Fix #151425: Outliner: Activating a camera clears the scene camera
This was only going wrong from the Outliner "Blender File" view.

b83b0d73a67a made it so that activating a Camera from the Outliner also
sets the scene active camera.
That might be fine, however in the Outliner "Blender File" view we
cannot use `outliner_search_back` to get back to the camera object
(there is simply no object up in the hierarchy in this case).

To resolve, only set the Scene camera in case we successfully found an
object via `outliner_search_back`.