Fix: Heap-buffer-overflow when dropping shape-keys
The `ShapeKeyDropTarget::on_drop()` got an array of selected shapekey
pointers from the drag custom-data, and iterated it until a null-pointer
was occured.

However when allocating the array, there was no null-pointer added to
the end of the vector. So it would keep iterating through memory until
hitting a null-pointer by chance.