UI: Default 'Save Modified Images' checkbox to enabled
Prior to this commit, the "Save Modified Images" checkbox when closing a
file would remember the previous value from when it was modified,
leading to unintuitive behavior (e.g. if the dialog box was
opened, modified and then the operation cancelled) and potential data
loss.

This commit moves the variable to the `wmWindowManagerRuntime` struct
instead of a static variable and resets the value when creating the
dialog box.

Part of #134518

# Conflicts:
#	source/blender/blenkernel/BKE_wm_runtime.hh