Fix #147809: VSE: Can't tweak handles in Industry Compatible keymap
Change `sequencer.select` keymap item to activate only on click, which
matches the default Blender keymap. Incidentally, this also fixes the
blade tool, which was broken (unreported).

We may not even need this keymap item for either keymap, but to avoid
any regressions it's best to just keep it for now.

Also add `use_restore_handle_selection` and `view2d_edge_pan` to the
Industry Compatible sequence slide items.