Vulkan: Synchronize features with KosmicKrisp
KosmicKrisp is a Vulkan wrapper on top of Metal. It has seen its first release
earlier this week. This PR updates the feature set with the release, to ensure
that more Blender functionality can be used.

This PR allows almost everything to work, except:
- Debugging overlays that developers uses (requires vertex atomics and stores)
- Some edit mesh overlays (requires provoking vertices)
- Wireframe shader node in EEVEE (require barycentric coordinates)

The goal is only to have a single environment where developers can work on Metal
and Vulkan at the same time. There are no plans to replace the Metal backend with
Vulkan, due to other restrictions.