Fix #151033: File Output crashes if a single value vector exists
The File Output node crashes if one of the inputs is a single value 3D
vector. That's because the code assume that the buffer would be 4D due
to hardware limitations, but this isn't the case for single values.
There was also a typo in the comparison where we didn't actually
consider the number of channels in the texture.