Fix #151406: Hydra Storm crash with generated image file
Add an extra null pointer check to avoid crashing. But the issue that
triggered this was not having a proper colorspace specified for image saving.
It doesn't make sense to copy image saving settings from the scene here either,
just set the format to PNG which is the default for new scenes.