Fix #151488: crash joining specific objects
Another issue exposed by 340f9d7ff3

In 2f59968361d8, we were dealing with objects that had no vertex groups
in `mesh.vertex_group_names` yet did have a `CD_MDEFORMVERT` custom data
layer. This time, we are dealing with objects having a somewhat
"invalid" index to a vertexgroup in `MDeformWeight` `def_nr`.
(so specifically, `def_nr` was 481208968 in the file from the report,
yet the object only had a single vertexgroup...)

As in 2f59968361d8, I'd say this shouldnt be a valid mesh really,
checking for operations that are treating `def_nr` badly would be good
to do.

In 4.5 (or actually, prior to 113d91aba8b3), there was code in
`BKE_object_defgroup_index_map_apply` that checked for `MDeformWeight`
`def_nr` wrt. being "valid" [in the range of the map], so that was fine
there, 340f9d7ff3 is not the real "cause" of the issue (it only exposed
it since selection order now made a difference), so 113d91aba8b3
is the real culprit.

To resolve, assign a "valid" `def_nr` (clamping it to the last valid
index in the map) which _might_ put it in a "wrong" vertexgroup in
cornercases, but this is similar to how code handled it prior afaict.

NOTE: I mention "similar" because there is still issues left to fix I
guess. Painting weights on such a mesh can result in `MDeformVert`
`totweight` being larger than the amount of vertexgroups we have, the
code in `BKE_object_defgroup_index_map_apply` (prior to 113d91aba8b3)
fixed this by reducing the `totweight` and reallocating the
`MDeformWeight` array. Not sure if the mesh joining code is really the
right place for this though or if this should be fixed at some other
place? The PR leaves this in a somewhat "invalid" state, we could pull
in more code from `BKE_object_defgroup_index_map_apply` if we wanted...