GHOST: Add --no-native-pixels support to Win32, X11 and Wayland
Blender by default uses the native pixel size for high resolution
displays on macOS. The command line argument --no-native-pixels can be
used to disable this behavior. It was so far ignored on Windows and
Linux.
This PR implements the command line argument for Win32, X11 and Wayland.
On these platforms, --no-native-pixels leads to the OS level UI scale to
be overwritten to 1, i.e., no UI scale being applied.
This can help viewport rendering benchmarks where the size of the 3D
viewport area should not be dependent on the OS level UI scale set by
the user.