Refactor: Compositor: Move static cache outside of the context
This patch refactors the compositor code to move the static cache
manager from the context to the caller code, leaving only a reference to
it. This is helpful to make the context purely an implementation class,
not one that holds cached data that needs to be kept around. Front-ends
should then have the responsibility to keep the cache manager around as
needed.