Fix #151618: Overlay: Attribute Viewer Instanced Pointclouds
Instanced pointclouds were drawn with the incorrect shader, leading to
crashes when used in Vulkan. This PR uses the instance_sub_ pass as that
matches the push constants being used.

However I'm not sure what the exact ideas are, the solution could also
be that we need to setup a different instance_point_cloud_sub.