Fix crash when inserting a single node into a node group.
Inserting a single node into a group will add new sockets to the group.
This invalidates the node group interface and requires an update of
dependent nodes, in particular group input/output nodes inside the group.

The node insert operator runs `update_structure_type_interface` which relies
on a match between the group input sockets and the tree interface. This does
not work when the group input nodes have not yet been updated and crashes.

To ensure the group input nodes are fully updated, run a full `BKE_ntree_update`,
which syncs the group inputs to the tree in `update_individual_nodes` before
invoking `update_structure_type_interface`.