Fix: Node group Insert/Separate operators lose animation data on nodes
When inserting or separating nodes using these operators animation data
may be lost (fcurves and drivers on node sockets and properties).
This happens when the group contains a node of the same name already.
In that case the inserted/separated node is renamed, but the animation
base path used is only the name of the new node. No animation data can
be found in the source tree for that name, so it doesn't get copied.

Fix by using the actual original name of the node in both operators.