Python: make the conditional HTTP downloader multiprocessing-safe
Make it safe for multiple Blender processes to download to the same file
path.

While these processes are downloading, they will all write to separate
temporary files. When done downloading, they will use an atomic move
(except on Windows) to move the temporary file to its final location.

AFAIK there is no simple API to use on Windows to rename a file when the
target filename already exists.

The last process to finish its download will "win", in the sense that
the final state of the file will be determined by the last-finished
download.