Cleanup: Compositor: Remove unused poll method exception
The compositor evaluator had an exception for the Render Layers and
Cryptomatte nodes when figuring out if nodes are supported because they
failed due to hacks in their poll methods. Those methods are now removed
and we can remove those exceptions here as well as they are no longer
needed.