Anim: make it explicit that 'Allow Negative Frames' can cause issues
In the Animation preferences, move the 'Allow Negative Frames' option to
a new 'Advanced' panel. This panel is closed by default, making the
option harder to find. Also an explanation is included that states

> Using negative frames can cause some issues, for example with audio
> playback or with some exporters.

This should make it clearer to useres why the option is not shown by
default. Before this PR, 'Allow Negative Frames' was shown as the very
first Animation option, and the fact that it could cause problems was
not communicated at all.

Related reports: #150608, #150610