GPU: Shader Tool:  Add builtin support for enum inside structs
This is an automatic replacement of the machinery we implemented
inside `eevee_light_shared.hh`. Instead, this generate the
appropriate code only for the shader implementation.

This auto generated code is safer and less error prone than
manually implemented accessors.

Moreover the accessor syntax is simplified to the function
call `()` operator.