WIP: Fix #151571: Grease Pencil render performance regression
The `DEG_OBJECT_ITER` in `DRW_render_check_grease_pencil`
is very expensive especially when there is instanced geometry that
needs to be expanded into temporary objects.

Instead, when an object is evaluated, we cache the contained
geometry types on its runtime data.

Then we can simply iterate over the objects in the
depsgraph and check if they are visible and if
they contain any Grease Pencil geometry.