Fix #147500: Separate Widths for Sidebar Category Tabs
Regions with panel categories can be a single width or hidden. So when
at the minimized size (showing only the category tabs) there isn't any
saved size to restore to, so it can only be changed to an arbitrary
width. This PR allows each category to store/restore their own width,
so they can all vary if you wish and always restore to the prior size.