UI: Improved File Overwrite Feedback
When using the File Browser to save a file we show a caution that the
file selected exists by showing the name input red. But it isn't always
obvious what this is, and it does not update as you alter the name.
This PR adds "redalert" state to the "save" button as well and also
makes it show as "Overwrite".  And file name collisions are detected
while editing the file name.