Fix #151629: VSE: Prefetching replace blend adjustment strip crashes
The problem is that adjustment strips normally dont filter out the
strips beneath them in `query_rendered_strips`, but when their blend
mode is 'replace' they do. This hides the scene strip from
`seq_prefetch_scene_strip_is_rendered`, so it appears as if the prefetch
thread can render the frame even though it can't.

Adjustment strips can be thought of as effects where all strips on
channels beneath them function as inputs (since they're indirectly
rendered by them), but technically, they don't have any real inputs.

Thus, the existing `relations_render_loop_check` is not enough and we
need a new condition. This could be potentially replaced in the future
with a `query_indirectly_rendered_strips` which should achieve the same
result, but for now, this simple fix solves the problem.