Vulkan: Disable host image copy for Windows/NVIDIA
When using host image copy on Windows/NVIDIA, the allocation of images
can fail when Blender allocates a texture with only
GPU_TEXTURE_USAGE_SHADER_READ/WRITE usage.

In this case the VMA cannot find a memory block for allocation. This
could be related to specific flags that should be added. Changing these
flags requires more testing and proof.

For now host image copy will be disabled when using Windows/NVIDIA until
we know what is expected by VMA and the driver.