Assets: add a C++ wrapper for the Disk File Hash Service
Add a C++ wrapper around the Disk File Hash Service introduced in
!151497.

This PR adds two new functions to the BPY C++ interface. Both run a
script with specific local variables; one of them can also return a
value to the C++ code.

```cpp
bool BPY_run_string_with_locals(
    bContext *C,
    const blender::StringRefNull script,
    IDProperty &locals);
```

This function:

- obtains the GIL and sets things up, like `bpy_run_string_impl()`,
- uses the `locals` group IDProperty to populate the local variables for
  the script,
- executes the script,
- handle raised exceptions, and
- perform cleanup.
- Returns an 'ok' bool, so on failure (like an exception in Python)
  returns false.

Part of this code was introduced for the Remote Assets Library project
in !145257. This PR takes the code that's not related to assets, and
targets the main branch.

```cpp
std::optional<IDProperty *> BPY_run_string_with_locals_return_idprop(
    bContext *C,
    blender::StringRefNull script,
    IDProperty &locals);
```

This function performs the same as above, and assumes that the script
assigns a value to the `_result` local variable. Its value is returned
as `IDProperty`, or `std::nullopt` when there was an issue (a Python
exception, nothing assigned to `_result`, or a complex object assigned
to `_result`). `None` is returned as `nullptr` value.

Note: To be able to test this code in isolation, I added two function
calls to the WindowManager RNA interface. This is **NOT** meant to land
in `main`. However, attempts at writing a C++ unit test that gets
Blender so far that it can run Python have failed.