UI: Python: Hide (most) core add-ons
This makes so that most of the core add-ons are not perceived as add-ons
by the users. The fact that some of those features were written as
add-ons is then simply "implementation detail".

It introduces a new `preferences.addons_ui` property to be used for
example for the workspace filtering/listing functionality.

This has been the plan since the Extensions project. But it was left as
technical debt.

Add-ons tackled here:
* Cycles
* BVH
* FBX
* SVG
* UV Layout
* Blender Extensions

Now those add-ons are not listed on the Add-ons tab on the Preferences,
nor on the Workspace Filter Add-ons list. The pkg (Blender Extensions)
was already not showing on the add-ons list, but now it doesn't show
elsewhere either.

Those are the simple cases that don't have preferences exclusively on
the add-ons editor. The next good candidate is the glTF to complete the
IO add-ons.

Those add-ons are also always "enabled" on startup, to make sure lite
builds or old Blender don't interfere with them being active.

Note: The moment we hide the gltf2 add-on as well, we can remove the
bl_owner_use_filter code.

Ref. #151646