Compositor: Use type default values for single inputs fallback
This patch changes the fallback values when an image is erroneously
connected to a single value input. The fallback values are now whatever
the default value of the CPP type is, which is either an identity or a
zero value. This paves the way for future refactoring of handing of
single value inputs.