Cycles: Do not tag all sockets for modified when only shadow catcher needs recalc
The previous logic of `flag & UPDATE_ALL` tags all sockets for modified
even when only some flags are tagged for update, so that we could not
trust `xxx_is_modified()`.

Instead, we only call `tag_modified()` when the flag is equal to
`UPDATE_ALL`. Now we need to deal with the case when `OBJECT_MANAGER` is
tagged for update, which, according to where it is called in
`object.cpp`, is for tagging integrator's shadow catcher for update.