Fix #151759: Sort Append Filter List Internationally
This report notes that some Append "Advanced Filter" items are not
ordered correctly among a list that seems to be almost completely
alphabetical. Notably "Collections" and "Freestyle linestyles". These
are wrong because their identifiers are quite different from their
names ("collection" versus "group").  But if these are sorted by name
then it needs to be the translated name. If we just do that then items
with accents will be at the end of the list, like French "Éclairages".
This PR sorts these using translated strings with an international
sort based on the current output language.