Fix #146953: Toggle Asset Shelf and Its Header Together
In this bug report the user selects "Asset Shelf" from "Region Toggle
Pie". This does a python "context_toggle" of "show_region_asset_shelf"
which only unhides it but not the asset shelf header. This just makes
it so that the update of "show_region_asset_shelf" toggles both regions
together.