Fix #146786: Tooltip Disabled Message About External Assets
Hovering over some metatdata inputs in the Asset Browser can show a
tooltip that explains that this data cannot be edited because it is
in an external library. The text is a bit confusing because it ends
in "can" without a period and so looks clipped. This PR just changes
this text to be more descriptive. "Asset metadata from external asset
libraries cannot be edited, only assets stored in the current file can"
changed to "Asset metadata from external libraries cannot be edited.
Only assets stored in the current file can be edited."