Refactor: Cycles: Remove use of RNA C++ API: object.cpp
This one is a little less trivial than the other similar changes
recently, because it migrates depsgraph iteration from the RNA API to
the C++ depsgraph API, which is a bit more unruly. Previously we passed
around `BL::DepsgraphObjectInstance`, now we pass `::Object` and
`::DEGObjectIterData`.

Another thing to note is the duplication of `ED_mesh_split_faces` into
`mesh_split_edges_for_corner_normals`. I did that because we shouldn't
call the editors modules from Cycles, and because the splitting logic
is much more related to how Cycles decided to upload normals than any
Blender-internal logic.

Part of #130989