Fix #143117: Duplicated object remembers slot name from source object
Erase the 'last used slot identifier' when duplicating a data-block, if
that data-block has no Action assigned. This ensures that the following
works:

- Animate `Cube`, and in the Action editor see its slot name (`Cube`)
- Unassign the Action
- Duplicate `Cube`, this creates `Cube.001`
- Assign `CubeAction` to `Cube.001`
- Insert a key
- See that the Action now also has a slot named `Cube.001` for
  `Cube.001`.

Note that the behaviour when `Cube` still has its Action assigned has
not changed. Depending on the preferences, the duplicate will either
reuse the same Action (and thus also the same slot), or also duplicate
the Action (and thus still keep using the same slot in that duplicate).