FIX: Ensure get_color_space_by_interop_id returns primary
The same interop ID can be assigned to multiple color spaces in a config
file. This change ensures that get_color_space_by_interop_id returns the
right one, which requires the interop ID to also be either the name of
the color space or an alias.

A boolean is_primary_interop_id_ is added to indicate this and only
such color spaces are returned from get_color_space_by_interop_id.