Fix #151940: Denoise node has a memory leak
The Denoise node has a memory leak in certain cases when the inputs are
pre-filtered and come directly from an image node. That's because and
error in the code where derived resources did not get freed if the
result referenced external data. This patch fixes that by freeing
derived resources in that case.