Cleanup: Cycles, sign conversion in motion interpolation
Almost all calls to transform_motion_array_interpolate() had the number
of steps stored as signed integer (due to KernelObject::num_tfm_steps
stored as signed integer), and the calculation inside of this function
was also done on signed integer. The only place where the number of
steps is unsigned is in Object::compute_bounds() where a vector size is
passed as the number of steps.