Refactor: Compositor: Move handling for cache manager canceling
This patch moves the special handing needed in the cache manager to
support canceling during execution. Instead of having the logic in the
cache manager and evaluator, it is moved to the render compositor code,
which is the only place that has canceling. This simplifies future
refactors in this area.