Fix #151997: Dopesheet Shape Key Editor not properly updating geometry
When the `Solo Active Shape Key` option is used, selecting a shapekey
naturally should update the geometry (and it does when doing it from
Shape Keys panel in the Properties Editor.

Doing this from the Dopesheet Shape Key Editor does set the objects
`shapenr` but it is lacking the proper update handling done in
`rna_Object_active_shape_update` [which is a little more involved esp.
in editmode].

To resolve we could copy what `rna_Object_active_shape_update` is doing,
or better yet, just call an RNA update on the objects `shapenr`.