GPU: Shader Tool: Add std140 linting
The std140 has a few extra requirements that is better
not enforce everywhere.

Structure not matching these requirements do not have the
associated `_host_shared_uniform_` alias which is then
used for uniform buffers.

This allow to avoid most of the NO_BUFFER_TYPE_LINTING.

Struct padding is only enforced for structures that are less
than 32 bytes as they are the only use case of non aligned
structures.