Fix: Missing compositor relations in pipeline depsgraph
Commit b0278b24868 introduced code which was building compositor nodes
for all pipeline depsgraphs (regardless of the R_DOCOMP flag), but it
did not do it for the relations builder.

While it works in simple cases, it could easily lead to issues when
some non-trivial drivers depending on objects exist. Additionally, in
the !152082 we are trying to add sanitization for the depsgraph state,
and such nodes/relations discrepancy does not pass checks. Happens
with the camera_stereo_panoramic.blend render test (open the file, and
press F12 to render).