Fix #151621: Avoid overlap of volume objects with shared geometry
We have `face_overlap_avoidance` for volumes, but it was per geometry,
so it doesn't solve the problem for objects that share the same
geometry.

Now apply the offset directly to the object transformation. The current
offset should be close to the previous one, assuming the VDB object has
a bbox size of 100 in each direction.