Fix #152114: Metastrip length increases when moved
In !150945, `handles_set` did not only replace
`strip_time_translate_handles` (which did no clamping), but also
`time_handles_frame_set` (which could erroneously clamp if the new right
handle was left of the current left handle).

This manifested in the original report as a "lengthening when moving the
strips around quickly" but an easier way to notice the behavior is to
simply move the metastrip far to the right and then cancel the
transform.

To fix, add a check inside `handles_set` to avoid any clamping. This
should also fix some unreported bugs with other areas that used the
function.