Refactor: Convert many C style casts to C++ style and id_cast
For improved type safety and to follow code style guidelines, this is a
largely automated conversion of code in source/ and intern/cycles/blender.

* Use const_cast and static_cast where possible and reintepret_cast if not.
* For arithmetic and enumeration types, use functional notation.
* Prefer blender::id_cast where possible. This required some small code
  tweaks to avoid doing this cast before the ID type has been checked. There
  are probably a few more like this lurking that will trigger asserts but not
  affect release builds.
* Change Object.data to be ID* rather than void* to take more advantage of
  id_cast checks.

Not all casts where converted, only the ones that were easy to automate.
Various modules in intern/ would also benefit but are not handled as part of
this.