WIP: Experiment with 'dynamic override of ID pointers' in Depsgraph.
This is a proof of concept using the existing
`ViewLayer.material_override` property as 'dynamic override' moc-up,
replacing the render-engine handling of this option by a real override
in depsgraph building and evaluation.

See also the Dynamic Overrides design task: #132565.

Note that while functional at a basic level, this PR is nowhere near
complete (animdata ID p0inters e.g. are not handled at all here when
building the depsgraph). It is not even clear yet if the approach
taken here is the best one.

Further more, this fits fine with 'full ID replacement' (which is part
of the initial goals of the project), but will need more complex
solutions to support 'partial' ID replacements.