Fix: Cycles HIP-RT MNEE failure when build with SDK 7
Noticed when doing local development on a machine with the SDK 7
installed: MNEE test produces wrong render result.

Worked around by demoting some functions from forceinline to inline.

Affected functions are not necessarily trivial, so it might be good
idea to do it for all platforms to help the compiler.