Gitea: Add CODEOWNERS file
The file allows to automatically assign reviewers based on the
file path matching rules.

Currently is rather rudimentary to get a good feeling about how
well this works in practice. If it works well, then we will
definitely ask more developers to add themselves in the file, or
even map the entire source tree.